move load_and_prepare function to utils.files_handler -> can be used in debug scripts now.
